Migrate UpdaterService
