+++ src/AzzyBot.Bot/Services/CronJobs/AzzyBotUpdateCheckJob.cs
@@ -2,14 +2,16 @@
 using System.Threading;
 using System.Threading.Tasks;
 
+using AzzyBot.Bot.Services.Interfaces;
+
 using NCronJob;
 
 namespace AzzyBot.Bot.Services.CronJobs;
 
-public sealed class AzzyBotUpdateCheckJob(DiscordBotService botService, UpdaterService updater) : IJob
+public sealed class AzzyBotUpdateCheckJob(DiscordBotService botService, IUpdaterService updater) : IJob
 {
     private readonly DiscordBotService _botService = botService;
-    private readonly UpdaterService _updater = updater;
+    private readonly IUpdaterService _updater = updater;
 
     public async Task RunAsync(IJobExecutionContext context, CancellationToken token)
     {
