+++ src/AzzyBot.Bot/Services/Interfaces/IUpdaterService.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace AzzyBot.Bot.Services.Interfaces;
+
+public interface IUpdaterService
+{
+    IReadOnlyDictionary<string, string> GitHubHeaders { get; }
+    Task CheckForAzzyUpdatesAsync();
+}
