+++ src/AzzyBot.Bot/Services/UpdaterService.cs
@@ -4,6 +4,7 @@ using System.Globalization;
 using System.Text.Json;
 using System.Threading.Tasks;
 
+using AzzyBot.Bot.Services.Interfaces;
 using AzzyBot.Bot.Settings;
 using AzzyBot.Bot.Utilities;
 using AzzyBot.Bot.Utilities.Records;
@@ -18,7 +19,7 @@ using Microsoft.Extensions.Options;
 
 namespace AzzyBot.Bot.Services;
 
-public sealed class UpdaterService(ILogger<UpdaterService> logger, IOptions<AzzyBotSettings> botSettings, IOptions<CoreUpdaterSettings> updaterSettings, IDbActions dbActions, DiscordBotService botService, WebRequestService webService)
+public sealed class UpdaterService(ILogger<UpdaterService> logger, IOptions<AzzyBotSettings> botSettings, IOptions<CoreUpdaterSettings> updaterSettings, IDbActions dbActions, DiscordBotService botService, WebRequestService webService) : IUpdaterService
 {
     private readonly ILogger<UpdaterService> _logger = logger;
     private readonly AzzyBotSettings _botSettings = botSettings.Value;
@@ -72,25 +73,6 @@ public sealed class UpdaterService(ILogger<UpdaterService> logger, IOptions<Azzy
         await SendUpdateMessageAsync(onlineVersion, releaseDate, updaterRecord.Body);
     }
 
-    public static bool CheckUpdateNotification(int notifyCounter, in DateTimeOffset lastNotificationTime)
-    {
-        DateTimeOffset now = DateTimeOffset.UtcNow;
-        if (notifyCounter < 3 && now - lastNotificationTime > TimeSpan.FromHours(23.98))
-        {
-            return true;
-        }
-        else if (notifyCounter >= 3 && now - lastNotificationTime > TimeSpan.FromHours(11.98))
-        {
-            return true;
-        }
-        else if (notifyCounter >= 7 && now - lastNotificationTime > TimeSpan.FromHours(5.98))
-        {
-            return true;
-        }
-
-        return false;
-    }
-
     private async Task SendUpdateMessageAsync(string updateVersion, DateTimeOffset releaseDate, string changelog)
     {
         if (_lastOnlineVersion != updateVersion)
@@ -99,7 +81,7 @@ public sealed class UpdaterService(ILogger<UpdaterService> logger, IOptions<Azzy
             _azzyNotifyCounter = 0;
         }
 
-        if (!CheckUpdateNotification(_azzyNotifyCounter, _lastAzzyUpdateNotificationTime))
+        if (!Misc.CheckUpdateNotification(_azzyNotifyCounter, _lastAzzyUpdateNotificationTime))
             return;
 
         _lastAzzyUpdateNotificationTime = DateTimeOffset.UtcNow;
